+++ mobile/android/android-components/samples/browser/src/geckoBeta/java/org/mozilla/samples/browser/Components.kt
@@ -6,7 +6,11 @@ package org.mozilla.samples.browser
 
 import android.content.Context
 import mozilla.components.browser.engine.gecko.GeckoEngine
+import mozilla.components.browser.session.Session
 import mozilla.components.concept.engine.Engine
+import mozilla.components.concept.engine.EngineSession
+import mozilla.components.concept.engine.webextension.WebExtension
+import mozilla.components.concept.engine.webextension.WebExtensionTabDelegate
 import mozilla.components.feature.webcompat.WebCompatFeature
 
 /**
@@ -16,6 +20,12 @@ class Components(applicationContext: Context) : DefaultComponents(applicationCon
     override val engine: Engine by lazy {
         GeckoEngine(applicationContext, engineSettings).also {
             WebCompatFeature.install(it)
+
+            it.registerWebExtensionTabDelegate(object : WebExtensionTabDelegate {
+                override fun onNewTab(webExtension: WebExtension?, url: String, engineSession: EngineSession) {
+                    sessionManager.add(Session(url), true, engineSession)
+                }
+            })
         }
     }
 }